backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Flask-CORS==5.0.1
PyMySQL==1.1.1
cryptography==44.0.0
python-dotenv==1.1.0
Werkzeug==3.1.3
